requirements.txt
requests>=2.31.0
python-telegram-bot>=20.0
python-dateutil>=2.8.0
pytz>=2023.3
